split start into multiple components
