+++ src/views/workflow-actions/workflow-action-start-retry-policy/workflow-action-start-retry-policy.types.ts
@@ -1,0 +1,3 @@
+import { type SubFormProps } from '../workflow-action-start-form/workflow-action-start-form.types';
+
+export type Props = SubFormProps;
